azure_ml_lora/requirements.txt
# Azure ML LoRA Training Pipeline Dependencies

# Core ML libraries
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
azure-core>=1.29.0

# HTTP client
httpx>=0.24.0

# Machine Learning
transformers>=4.35.0
peft>=0.4.0
datasets>=2.14.0
torch>=2.0.0
accelerate>=0.20.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Development
pytest>=7.0.0
black>=23.0.0